Replace RGB and RGBA iterators and remove `Clone` bounds
Rewrite the `iter` methods on `RGB<T>` and `RGBA<T>` to return a custom
struct in typical Rust fashion rather than casting to a slice and then
cloning. The returned iterators are over references instead of owned
values, in keeping with the usual meaning of `iter`. In the common case
where `T: Copy`, one just needs to add `.copied()` to the method chain.